+++ qc-application/src/main/java/com/eager/questioncloud/application/api/user/register/RegisterUserController.java
@@ -4,8 +4,8 @@ import com.eager.questioncloud.application.api.common.DefaultResponse;
 import com.eager.questioncloud.application.api.user.register.RegisterUserControllerRequest.CreateUserRequest;
 import com.eager.questioncloud.application.api.user.register.RegisterUserControllerResponse.CreateUserResponse;
 import com.eager.questioncloud.core.domain.user.model.User;
-import com.eager.questioncloud.core.domain.verification.EmailVerification;
-import com.eager.questioncloud.core.domain.verification.EmailVerificationType;
+import com.eager.questioncloud.core.domain.verification.enums.EmailVerificationType;
+import com.eager.questioncloud.core.domain.verification.model.EmailVerification;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import jakarta.validation.Valid;
