+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/verification/enums/EmailVerificationType.java
@@ -1,4 +1,4 @@
-package com.eager.questioncloud.core.domain.verification;
+package com.eager.questioncloud.core.domain.verification.enums;
 
 import lombok.Getter;
