+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/verification/implement/EmailVerificationProcessor.java
@@ -1,5 +1,8 @@
-package com.eager.questioncloud.core.domain.verification;
+package com.eager.questioncloud.core.domain.verification.implement;
 
+import com.eager.questioncloud.core.domain.verification.enums.EmailVerificationType;
+import com.eager.questioncloud.core.domain.verification.infrastructure.EmailVerificationRepository;
+import com.eager.questioncloud.core.domain.verification.model.EmailVerification;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
